Command-not-found also works for zsh now
This is a patch for `cnf` package.

`10-command-not-found.sh` is located in `/usr/share/defaults/etc/profile.d`,
and is sourced by start files of `bash`.

As I intend to unifying startup files of both `bash` and `zsh`, it would
also be sourced by `zsh`, if my suggestion is accepted.

It works for `zsh` now with the following modificatiosn:

- This functionality shall only be enabled for interactive shell, and is
  tested by `[ "$(expr $- : '.*i')" -ne 0 ]`.
  Though this test expression looks ugly, it is POSIX compatible and doesn't
  require external program.

- I replaced the obsolescent ` `foo bar` ` syntax with ` $(foo bar) `.

- The test of UID was `[ "`id -u`" == "0" ]`, where `==` is not
  Bourne-compatible and doesn't work for `zsh` unless I replace the
  `[` with `[[`. But `[[` itself is also not Bourne-compatible. So I
  decide to have `[ "$(id -u)" = "0" ]` instead, which works for
  Bourne-compatible shells.

- The `command_not_found_handle` is loaded as long as it's interactive
  shell, and is used directly by `bash`. When
  it's `zsh` shell, then a wrapper function `command_not_found_handler`
  is defined.